go/oasis-node: Move storage benchmark subcommand under debug
