+++ .changelog/2832.internal.1.md
@@ -1,0 +1,1 @@
+go/oasis-node: Move storage benchmark subcommand under debug
