--- go/oasis-node/cmd/storage/storage.go
@@ -1,20 +1,0 @@
-// Package storage implements various storage related sub-commands.
-package storage
-
-import (
-	"github.com/spf13/cobra"
-
-	"github.com/oasislabs/oasis-core/go/oasis-node/cmd/storage/benchmark"
-)
-
-var storageCmd = &cobra.Command{
-	Use:   "storage",
-	Short: "storage services and utilities",
-}
-
-// Register registers the storage sub-command and all of it's children.
-func Register(parentCmd *cobra.Command) {
-	benchmark.Register(storageCmd)
-
-	parentCmd.AddCommand(storageCmd)
-}
